Refactor Sponsor route
